feat: updated artifacts with no-zip option
