+++ Tests/RunnerPart2/Runner2Test.php
@@ -15,9 +15,6 @@ use Keboola\StorageApi\Client as StorageApiClient;
 use Keboola\StorageApi\Components;
 use Keboola\StorageApi\Options\Components\Configuration;
 use Keboola\StorageApi\Options\Components\ListComponentConfigurationsOptions;
-use Keboola\StorageApi\Options\FileUploadOptions;
-use Keboola\StorageApi\Options\ListFilesOptions;
-use Symfony\Component\Process\Process;
 use Throwable;
 
 class Runner2Test extends BaseRunnerTest
@@ -274,433 +271,5 @@ class Runner2Test extends BaseRunnerTest
         // check the component ran but no connection string was passed
         self::assertContains('Environment "KBC_PROJECTID" has value "1234".', $containerOutput);
         self::assertNotContains('Environment "AZURE_STORAGE_CONNECTION_STRING"', $containerOutput);
-    }
-
-    public function testArtifactsUpload()
-    {
-        $storageApiMock = $this->getMockBuilder(StorageApiClient::class)
-            ->setConstructorArgs([[
-                'url' => STORAGE_API_URL,
-                'token' => STORAGE_API_TOKEN,
-            ]])
-            ->onlyMethods(['verifyToken', 'listFiles'])
-            ->getMock()
-        ;
-        $storageApiMock->method('verifyToken')->willReturn([
-            'owner' => [
-                'id' => '1234',
-                'fileStorageProvider' => 'local',
-                'features' => ['artifacts'],
-            ],
-        ]);
-        $storageApiMock->method('listFiles')->willReturn([]);
-        $this->setClientMock($storageApiMock);
-
-        $componentData = [
-            'id' => 'keboola.python-transformation',
-            'data' => [
-                'definition' => [
-                    'type' => 'aws-ecr',
-                    'uri' => '147946154733.dkr.ecr.us-east-1.amazonaws.com/developer-portal-v2/keboola.python-transformation',
-                ],
-            ],
-        ];
-
-        $config = [
-            'storage' => [],
-            'parameters' => [
-                'script' => [
-                    'import os',
-                    'path = "/data/artifacts/out/current"',
-                    'if not os.path.exists(path):os.makedirs(path)',
-                    'pathShared = "/data/artifacts/out/shared"',
-                    'if not os.path.exists(path):os.makedirs(pathShared)',
-                    'with open("/data/artifacts/out/current/myartifact1", "w") as file:',
-                    '   file.write("value1")',
-                    'with open("/data/artifacts/out/shared/myartifact1", "w") as file:',
-                    '   file.write("value1")',
-                ],
-            ],
-        ];
-
-        $components = new Components($storageApiMock);
-        $configuration = $components->addConfiguration(
-            (new Configuration())
-                ->setComponentId('keboola.python-transformation')
-                ->setConfiguration($config)
-                ->setName('artifacts tests')
-        );
-
-        $configId = $configuration['id'];
-        $jobId = (string) random_int(0, 999999);
-        $orchestrationId = (string) random_int(0, 999999);
-
-        $runner = $this->getRunner();
-        $outputs = [];
-        $runner->run(
-            $this->prepareJobDefinitions(
-                $componentData,
-                $configId,
-                $config,
-                []
-            ),
-            'run',
-            'run',
-            $jobId,
-            new NullUsageFile(),
-            [],
-            $outputs,
-            null,
-            $orchestrationId
-        );
-
-        sleep(2);
-
-        // current
-        $files = $this->client->listFiles(
-            (new ListFilesOptions())
-                ->setQuery(sprintf(
-                    'tags:(artifact AND branchId-default AND componentId-keboola.python-transformation '
-                    . 'AND configId-%s AND jobId-%s NOT shared)',
-                    $configId,
-                    $jobId
-                ))
-                ->setLimit(1)
-        );
-
-        $currentFile = $files[0];
-        self::assertEquals('artifacts.tar.gz', $currentFile['name']);
-        self::assertContains('branchId-default', $currentFile['tags']);
-        self::assertContains('componentId-keboola.python-transformation', $currentFile['tags']);
-        self::assertContains('configId-' . $configId, $currentFile['tags']);
-        self::assertContains('jobId-' . $jobId, $currentFile['tags']);
-
-        // shared
-        $files = $this->client->listFiles(
-            (new ListFilesOptions())
-                ->setQuery(sprintf(
-                    'tags:(artifact AND shared AND branchId-default '
-                    . 'AND componentId-keboola.python-transformation AND orchestrationId-%s)',
-                    $orchestrationId
-                ))
-                ->setLimit(1)
-        );
-
-        $sharedFile = $files[0];
-        self::assertEquals('artifacts.tar.gz', $sharedFile['name']);
-        self::assertContains('branchId-default', $sharedFile['tags']);
-        self::assertContains('componentId-keboola.python-transformation', $sharedFile['tags']);
-        self::assertContains('configId-' . $configId, $sharedFile['tags']);
-        self::assertContains('jobId-' . $jobId, $sharedFile['tags']);
-        self::assertContains('orchestrationId-' . $orchestrationId, $sharedFile['tags']);
-        self::assertContains('shared', $sharedFile['tags']);
-
-        /** @var Output $output */
-        $output = $outputs[0];
-        self::assertSame(
-            [
-                [
-                    'storageFileId' => $currentFile['id'],
-                ],
-                [
-                    'storageFileId' => $sharedFile['id'],
-                ],
-            ],
-            $output->getArtifactsUploaded()
-        );
-    }
-
-    public function testArtifactsUploadEmpty()
-    {
-        $storageApiMock = $this->getMockBuilder(StorageApiClient::class)
-            ->setConstructorArgs([[
-                'url' => STORAGE_API_URL,
-                'token' => STORAGE_API_TOKEN,
-            ]])
-            ->onlyMethods(['verifyToken', 'listFiles'])
-            ->getMock()
-        ;
-        $storageApiMock->method('verifyToken')->willReturn([
-            'owner' => [
-                'id' => '1234',
-                'fileStorageProvider' => 'local',
-                'features' => ['artifacts'],
-            ],
-        ]);
-        $storageApiMock->method('listFiles')->willReturn([]);
-        $this->setClientMock($storageApiMock);
-
-        $componentData = [
-            'id' => 'keboola.python-transformation',
-            'data' => [
-                'definition' => [
-                    'type' => 'aws-ecr',
-                    'uri' => '147946154733.dkr.ecr.us-east-1.amazonaws.com/developer-portal-v2/keboola.python-transformation',
-                ],
-            ],
-        ];
-
-        $config = [
-            'storage' => [],
-            'parameters' => [
-                'script' => [
-                    'import os',
-                    '# do nothing'
-                ],
-            ],
-        ];
-
-        $components = new Components($storageApiMock);
-        $configuration = $components->addConfiguration(
-            (new Configuration())
-                ->setComponentId('keboola.python-transformation')
-                ->setConfiguration($config)
-                ->setName('artifacts tests')
-        );
-
-        $configId = $configuration['id'];
-        $jobId = (string) random_int(0, 999999);
-
-        $runner = $this->getRunner();
-        $outputs = [];
-        $runner->run(
-            $this->prepareJobDefinitions(
-                $componentData,
-                $configId,
-                $config,
-                []
-            ),
-            'run',
-            'run',
-            $jobId,
-            new NullUsageFile(),
-            [],
-            $outputs,
-            null
-        );
-
-        sleep(2);
-
-        $files = $this->client->listFiles(
-            (new ListFilesOptions())
-                ->setQuery(sprintf(
-                    'tags:(artifact AND branchId-default AND componentId-keboola.python-transformation '
-                    . 'AND configId-%s AND jobId-%s)',
-                    $configId,
-                    $jobId
-                ))
-                ->setLimit(1)
-        );
-        self::assertEmpty($files);
-
-        /** @var Output $output */
-        $output = $outputs[0];
-        self::assertSame([], $output->getArtifactsUploaded());
-    }
-
-    public function testArtifactsDownload()
-    {
-        $storageApiMock = $this->getMockBuilder(StorageApiClient::class)
-            ->setConstructorArgs([[
-                'url' => STORAGE_API_URL,
-                'token' => STORAGE_API_TOKEN,
-            ]])
-            ->onlyMethods(['verifyToken'])
-            ->getMock()
-        ;
-        $storageApiMock->method('verifyToken')->willReturn([
-            'owner' => [
-                'id' => '1234',
-                'fileStorageProvider' => 'local',
-                'features' => ['artifacts'],
-            ],
-        ]);
-        $this->setClientMock($storageApiMock);
-
-        $previousJobId = rand(0, 999999);
-        $config = [
-            'storage' => [],
-            'parameters' => [
-                'script' => [
-                    'import os',
-                    sprintf('with open("/data/artifacts/in/runs/jobId-%s/artifact1", "r") as f:', $previousJobId),
-                    '   print(f.read())',
-                ],
-            ],
-            'artifacts' => [
-                'runs' => [
-                    'enabled' => true,
-                    'filter' => [
-                        'limit' => 1,
-                    ],
-                ],
-            ],
-        ];
-        $components = new Components($storageApiMock);
-        $configuration = $components->addConfiguration(
-            (new Configuration())
-                ->setComponentId('keboola.python-transformation')
-                ->setConfiguration($config)
-                ->setName('artifacts tests')
-        );
-        $configId = $configuration['id'];
-
-        if (!is_dir('/tmp/artifact/')) {
-            mkdir('/tmp/artifact/');
-        }
-        file_put_contents('/tmp/artifact/artifact1', 'value1');
-
-        $process = new Process([
-            'tar',
-            '-C',
-            '/tmp/artifact',
-            '-czvf',
-            '/tmp/artifacts.tar.gz',
-            '.',
-        ]);
-        $process->mustRun();
-
-        $uploadedFileId = $this->client->uploadFile(
-            '/tmp/artifacts.tar.gz',
-            (new FileUploadOptions())
-                ->setTags([
-                    'artifact',
-                    'branchId-default',
-                    'componentId-keboola.python-transformation',
-                    'configId-' . $configId,
-                    'jobId-' . $previousJobId,
-                ])
-        );
-
-        $componentData = [
-            'id' => 'keboola.python-transformation',
-            'data' => [
-                'definition' => [
-                    'type' => 'aws-ecr',
-                    'uri' => '147946154733.dkr.ecr.us-east-1.amazonaws.com/developer-portal-v2/keboola.python-transformation',
-                ],
-            ],
-        ];
-
-        sleep(2);
-
-        $runner = $this->getRunner();
-        $outputs = [];
-        $runner->run(
-            $this->prepareJobDefinitions(
-                $componentData,
-                $configId,
-                $config,
-                []
-            ),
-            'run',
-            'run',
-            '1234567',
-            new NullUsageFile(),
-            [],
-            $outputs,
-            null
-        );
-
-        /** @var Output $output */
-        $output = $outputs[0];
-        self::assertStringContainsString('value1', $output->getProcessOutput());
-        self::assertSame([
-            ['storageFileId' => $uploadedFileId],
-        ], $output->getArtifactsDownloaded());
-    }
-
-    public function testArtifactsDownloadEmpty()
-    {
-        $storageApiMock = $this->getMockBuilder(StorageApiClient::class)
-            ->setConstructorArgs([[
-                'url' => STORAGE_API_URL,
-                'token' => STORAGE_API_TOKEN,
-            ]])
-            ->onlyMethods(['verifyToken'])
-            ->getMock()
-        ;
-        $storageApiMock->method('verifyToken')->willReturn([
-            'owner' => [
-                'id' => '1234',
-                'fileStorageProvider' => 'local',
-                'features' => ['artifacts'],
-            ],
-        ]);
-        $this->setClientMock($storageApiMock);
-
-        $config = [
-            'storage' => [],
-            'parameters' => [
-                'script' => [
-                    'import os',
-                    '# do nothing'
-                ],
-            ],
-            'artifacts' => [
-                'runs' => [
-                    'enabled' => false,
-                ],
-            ],
-        ];
-        $components = new Components($storageApiMock);
-        $configuration = $components->addConfiguration(
-            (new Configuration())
-                ->setComponentId('keboola.python-transformation')
-                ->setConfiguration($config)
-                ->setName('artifacts tests')
-        );
-        $configId = $configuration['id'];
-
-        if (!is_dir('/tmp/artifact/')) {
-            mkdir('/tmp/artifact/');
-        }
-
-        $componentData = [
-            'id' => 'keboola.python-transformation',
-            'data' => [
-                'definition' => [
-                    'type' => 'aws-ecr',
-                    'uri' => '147946154733.dkr.ecr.us-east-1.amazonaws.com/developer-portal-v2/keboola.python-transformation',
-                ],
-            ],
-        ];
-
-        sleep(2);
-
-        $runner = $this->getRunner();
-        $outputs = [];
-        $runner->run(
-            $this->prepareJobDefinitions(
-                $componentData,
-                $configId,
-                $config,
-                []
-            ),
-            'run',
-            'run',
-            '1234567',
-            new NullUsageFile(),
-            [],
-            $outputs,
-            null
-        );
-
-        /** @var Output $output */
-        $output = $outputs[0];
-        self::assertSame([], $output->getArtifactsDownloaded());
-    }
-
-    /**
-     * @param array $componentData
-     * @param $configId
-     * @param array $configData
-     * @param array $state
-     * @return JobDefinition[]
-     */
-    private function prepareJobDefinitions(array $componentData, $configId, array $configData, array $state)
-    {
-        $jobDefinition = new JobDefinition($configData, new Component($componentData), $configId, 'v123', $state);
-        return [$jobDefinition];
     }
 }
