+++ Tests/backend-specific-tests/Exasol/RunnerExasolTest.php
@@ -79,19 +79,6 @@ class RunnerExasolTest extends BaseRunnerTest
         $this->getClient()->createBucket('exasol-runner-test', Client::STAGE_OUT, 'Docker TestSuite', 'exasol');
     }
 
-    /**
-     * @param array $componentData
-     * @param $configId
-     * @param array $configData
-     * @param array $state
-     * @return JobDefinition[]
-     */
-    protected function prepareJobDefinitions(array $componentData, $configId, array $configData, array $state)
-    {
-        $jobDefinition = new JobDefinition($configData, new Component($componentData), $configId, 'v123', $state);
-        return [$jobDefinition];
-    }
-
     public function testWorkspaceExasolMapping()
     {
         $this->clearBuckets();
