+++ Tests/backend-specific-tests/Synapse/RunnerSynapseTest.php
@@ -90,19 +90,6 @@ class RunnerSynapseTest extends BaseRunnerTest
         $this->getClient()->createBucket('synapse-runner-test', Client::STAGE_OUT, 'Docker TestSuite', 'synapse');
     }
 
-    /**
-     * @param array $componentData
-     * @param $configId
-     * @param array $configData
-     * @param array $state
-     * @return JobDefinition[]
-     */
-    protected function prepareJobDefinitions(array $componentData, $configId, array $configData, array $state)
-    {
-        $jobDefinition = new JobDefinition($configData, new Component($componentData), $configId, 'v123', $state);
-        return [$jobDefinition];
-    }
-
     public function testWorkspaceSynapseMapping()
     {
         $this->clearBuckets();
