+++ composer.json
@@ -12,7 +12,7 @@
         "php": ">=7.4",
         "ext-json": "*",
         "ext-zip": "*",
-        "keboola/artifacts": "^0.2.0",
+        "keboola/artifacts": "^0.3.0",
         "keboola/common-exceptions": "^1.0",
         "keboola/gelf-server": "^3.0",
         "keboola/input-mapping": "^16.1",
